requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==4.1.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
flake8==5.0.4
mccabe==0.7.0
packaging==21.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytz==2022.5
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
wrapt==1.14.1
